Add files via upload
Hand gesture recognition has emerged as a pivotal area of research within computer 
vision, fuelled by its potential to redefine human-computer interaction. The primary aim 
of this research is to contribute to this evolving field by developing a real-time hand 
gesture recognition system capable of accurately interpreting and responding to a diverse 
range of hand movements. Leveraging the MediaPipe and OpenCV libraries, this system 
seeks to overcome existing limitations in gesture recognition technology, particularly in 
terms of accuracy and real-time performance. By harnessing deep learning techniques 
and advanced computer vision algorithms, the system aims to provide users with an 
intuitive means of interacting with digital devices, ultimately enhancing the overall user 
experience.
In addition to enhancing human-computer interaction, this research also aims to facilitate 
applications such as sign language translation, where precise interpretation of hand 
gestures is crucial. By developing a robust hand gesture recognition system, this research 
seeks to bridge the gap between digital devices and users who rely on sign language as 
their primary mode of communication. Furthermore, by ensuring the system's 
adaptability to diverse environmental conditions, including varying lighting conditions 
and background clutter, this research aims to broaden the practical applicability of 
gesture-based interaction technology. Ultimately, the goal is to empower individuals with 
diverse communication needs and provide them with accessible and inclusive means of 
engaging with digital technology.
Beyond its immediate applications, the development of a real-time hand gesture 
recognition system holds broader implications for the future of human-computer 
interaction and technological innovation. By pushing the boundaries of what is possible 
in gesture recognition technology, this research opens new avenues for creativity and 
exploration in fields such as virtual reality, augmented reality, robotics, and assistive 
technologies. Moreover, by contributing to the ongoing advancements in gesture-based 
interaction systems, this research aims to inspire further innovation and collaboration 
within the research community. Ultimately, the impact of this research extends beyond 
the development of a single system, serving as a catalyst for continued progress and 
discovery in the dynamic field of computer vision and human-computer interaction.